refactor: restructure `api/` folder inner files
