+++ api/src/spawner/trpc.ts
@@ -8,7 +8,7 @@ import WebSocket from "ws";
 import { z } from "zod";
 
 // import { WebsocketProvider } from "../../ui/src/lib/y-websocket";
-import { WebsocketProvider } from "./y-websocket";
+import { WebsocketProvider } from "../yjs/y-websocket";
 
 import { killRuntime, spawnRuntime } from "./spawner_native";
