+++ ui/src/lib/trpc.ts
@@ -1,3 +1,3 @@
 import { createTRPCReact } from "@trpc/react-query";
-import type { AppRouter } from "../../../api/src/trpc";
+import type { AppRouter } from "../../../api/src/spawner/trpc";
 export const trpc = createTRPCReact<AppRouter>();
